@@ -1,22 +1,21 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
+  <Import Project="../core.props" />
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace></RootNamespace>
     <NoWarn>CS0649;CS0169</NoWarn>
     <NukeRootDirectory>..</NukeRootDirectory>
     <NukeScriptDirectory>..</NukeScriptDirectory>
+    <NukeTelemetryVersion>1</NukeTelemetryVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Nuke.Common" Version="5.0.2" />
-    <PackageReference Include="Candoumbe.MiscUtilities" Version="0.1.0" />
-    <PackageReference Include="ReportGenerator" Version="4.8.*" />
-    <PackageReference Include="GitVersion.CommandLine" Version="5.5.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
+    <PackageReference Include="Candoumbe.Pipelines" Version="0.12.1" />
+    <PackageDownload Include="ReportGenerator" Version="[5.1.12]" />
+    <PackageDownload Include="GitVersion.Tool" Version="[5.8.1]" />
+    <PackageDownload Include="Codecov.Tool" Version="[1.12.3]" />
+    <PackageReference Include="CodecovUploader" Version="0.8.0"/>
   </ItemGroup>
-
 </Project>
