@@ -1,14 +1,14 @@
-﻿using System;
+﻿using Optional;
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 
-using Optional;
+using static Candoumbe.Forms.FormFieldType;
 
-using static Forms.FormFieldType;
-
-namespace Forms
+namespace Candoumbe.Forms
 {
     /// <summary>
     /// Helper class to build a <see cref="Form"/> instance
@@ -16,32 +16,36 @@ namespace Forms
     /// <typeparam name="T">Type to build a <see cref="Form"/> for.</typeparam>
     public class FormBuilder<T>
     {
-        private static readonly HashSet<Type> _dateTypes = new HashSet<Type>
-        {
+        private static readonly HashSet<Type> DateTimeTypes =
+        [
             typeof(DateTime), typeof(DateTime?),
-            typeof(DateTimeOffset), typeof(DateTimeOffset?),
-        };
+            typeof(DateTimeOffset), typeof(DateTimeOffset?)
+        ];
 
-        private static readonly HashSet<Type> _numericTypes = new HashSet<Type>
-        {
+#if NET6_0_OR_GREATER
+        private static readonly HashSet<Type> DateTypes = [ typeof(DateOnly), typeof(DateOnly?) ];
+#endif
+
+        private static readonly HashSet<Type> NumericTypes =
+        [
             typeof(int), typeof(int?),
             typeof(float), typeof(float?),
             typeof(long), typeof(long?),
             typeof(double), typeof(double?),
             typeof(short), typeof(short?),
             typeof(decimal), typeof(decimal?)
-        };
+        ];
 
-        private readonly IList<FormField> _fields;
+        private readonly List<FormField> _fields;
         private readonly Link _meta;
 
         /// <summary>
         /// Creates a new <see cref="FormBuilder{T}"/> instance
         /// </summary>
-        /// <param name="meta">describes where and how to send the form's data</param>
+        /// <param name="meta">describes where and how to send the form's data.</param>
         public FormBuilder(Link meta = null)
         {
-            _fields = new List<FormField>();
+            _fields = [];
             _meta = meta;
         }
 
@@ -49,23 +53,30 @@ namespace Forms
         /// Adds a field to the <see cref="Form"/>'s configuration.
         /// </summary>
         /// <remarks>
-        /// 
         /// </remarks>
-        /// <typeparam name="TProperty"></typeparam>
-        /// <param name="property"></param>
+        /// <typeparam name="TProperty">The type of the property for which a <see cref="FormField"/> will be added.</typeparam>
+        /// <param name="property">The property of <typeparamref name="T"/> for which a field will be added to the configuration.</param>
         /// <param name="attributes">Overrides field's attributes</param>
-        /// <returns></returns>
+        /// <returns>The current instance.</returns>
         public FormBuilder<T> AddField<TProperty>(Expression<Func<T, TProperty>> property, FormFieldAttributeOverrides attributes = null)
         {
             if (property.Body is MemberExpression me)
             {
-                FormField field = new FormField { Name = me.Member.Name };
+                FormField field = new() { Name = me.Member.Name };
                 Option<FormFieldAttribute> optionalFormFieldAttribute = me.Member.GetCustomAttribute<FormFieldAttribute>()
                     .SomeNotNull();
-                UpdateAttributesField(property, field, optionalFormFieldAttribute, attributes.SomeNotNull());
+                field = UpdateAttributesField(property, field, optionalFormFieldAttribute, attributes.SomeNotNull());
 
+#if NET5_0_OR_GREATER
+                field = field with
+                {
+                    Label = field.Name,
+                    Enabled = attributes?.Enabled
+                };
+#else
                 field.Label = field.Name;
                 field.Enabled = attributes?.Enabled;
+#endif
 
                 _fields.Add(field);
             }
@@ -73,16 +84,22 @@ namespace Forms
             return this;
         }
 
-        private static void UpdateAttributesField<TProperty>(Expression<Func<T, TProperty>> property,
-                                                             FormField field,
-                                                             Option<FormFieldAttribute> optionalFormFieldAttribute,
-                                                             Option<FormFieldAttributeOverrides> optionalAttributesOverride)
+        private static FormField UpdateAttributesField<TProperty>(Expression<Func<T, TProperty>> property,
+                                                                 FormField field,
+                                                                 Option<FormFieldAttribute> optionalFormFieldAttribute,
+                                                                 Option<FormFieldAttributeOverrides> optionalAttributesOverride)
         {
-            if (_dateTypes.Contains(property.ReturnType))
+            if (DateTimeTypes.Contains(property.ReturnType))
             {
                 field.Type = FormFieldType.DateTime;
             }
-            else if (_numericTypes.Contains(property.ReturnType))
+#if NET6_0_OR_GREATER
+            else if (DateTypes.Contains(property.ReturnType))
+            {
+                field.Type = Date;
+            }
+#endif
+            else if (NumericTypes.Contains(property.ReturnType))
             {
                 field.Type = Integer;
             }
@@ -92,80 +109,157 @@ namespace Forms
                 {
                     if (attr.IsDescriptionSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Description = attr.Description };
+#else
                         field.Description = attr.Description;
+#endif
                     }
                     if (attr.IsSecretSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Secret = attr.Secret };
+#else
                         field.Secret = attr.Secret;
+#endif
                     }
 
                     if (attr.IsMinSet)
                     {
-                        field.Min = attr.Min;
+#if NET5_0_OR_GREATER
+                        field = field with { MinSize = attr.MinSize };
+#else
+                        field.MinSize = attr.MinSize;
+#endif
                     }
 
+#if NET5_0_OR_GREATER
+                    field = field with { Pattern = attr.Pattern };
+#else
                     field.Pattern = attr.Pattern;
+#endif
+
                     if (attr.IsTypeSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Type = attr.Type };
+#else
                         field.Type = attr.Type;
+#endif
                     }
+
                     if (attr.IsMinSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Min = attr.Min };
+#else
                         field.Min = attr.Min;
+#endif
                     }
 
                     if (attr.IsMaxSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Max = attr.Max };
+#else
                         field.Max = attr.Max;
+#endif
                     }
 
                     if (attr.IsMinLengthSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { MinLength = attr.MinLength };
+#else
                         field.MinLength = attr.MinLength;
+#endif
                     }
 
                     if (attr.IsMaxLengthSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { MaxLength = attr.MaxLength };
+#else
                         field.MaxLength = attr.MaxLength;
+#endif
                     }
+
                     if (attr.IsTypeSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Type = attr.Type };
+#else
                         field.Type = attr.Type;
+#endif
                     }
 
                     if (attr.IsMaxSizeSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { MaxSize = attr.MaxSize };
+#else
                         field.MaxSize = attr.MaxSize;
+#endif
                     }
 
                     if (attr.IsEnabledSet)
                     {
+#if NET5_0_OR_GREATER
+                        field = field with { Enabled = attr.Enabled };
+#else
                         field.Enabled = attr.Enabled;
+#endif
                     }
                 });
 
-            optionalAttributesOverride.MatchSome((attrs) =>
+            optionalAttributesOverride.MatchSome((attr) =>
             {
-                if (attrs.Min.HasValue)
+                if (attr.Min.HasValue)
                 {
-                    field.Min = attrs.Min;
+#if NET5_0_OR_GREATER
+                    field = field with { Min = attr.Min };
+#else
+                    field.Min = attr.Min;
+#endif
                 }
-                if (attrs.Secret.HasValue)
+                if (attr.Secret.HasValue)
                 {
-                    field.Secret = attrs.Secret;
+#if NET5_0_OR_GREATER
+                    field = field with { Secret = attr.Secret };
+#else
+                    field.Secret = attr.Secret;
+#endif
                 }
-                if (attrs.IsDescriptionSet)
+                if (attr.IsDescriptionSet)
                 {
-                    field.Description = attrs.Description;
+#if NET5_0_OR_GREATER
+                    field = field with { Description = attr.Description };
+#else
+                    field.Description = attr.Description;
+#endif
                 }
-                field.Label = attrs.Label ?? field.Name;
-                if (attrs.Max.HasValue)
+
+#if NET5_0_OR_GREATER
+                field = field with { Label = attr.Label ?? field.Name };
+#else
+                field.Label = attr.Label ?? field.Name;
+#endif
+
+                if (attr.Max.HasValue)
                 {
-                    field.Max = attrs.Max;
+#if NET5_0_OR_GREATER
+                    field = field with { Max = attr.Max };
+#else
+                    field.Max = attr.Max;
+#endif
                 }
-                if (attrs.Pattern != null)
+                if (attr.Pattern is not null)
                 {
-                    field.Pattern = attrs.Pattern;
+#if NET5_0_OR_GREATER
+                    field = field with { Pattern = attr.Pattern };
+#else
+                    field.Pattern = attr.Pattern;
+#endif
                 }
 
                 // if (attrs.IsMinLengthSet)
@@ -173,36 +267,68 @@ namespace Forms
                 //     field.MaxLength = attrs.MaxLength;
                 // }
 
-                if (attrs.IsMaxLengthSet)
+                if (attr.IsMaxLengthSet)
                 {
-                    field.MaxLength = attrs.MaxLength;
-                }
-
-                if (attrs.IsTypeSet)
-                {
-                    field.Type = attrs.Type;
+#if NET5_0_OR_GREATER
+                    field = field with { MaxLength = attr.MaxLength };
+#else
+                    field.MaxLength = attr.MaxLength;
+#endif
                 }
             });
+
+            return field;
         }
 
+        /// <summary>
+        /// Ands a <see cref="FormField"/> with several <see cref="FormFieldOption"/>s to the <see cref="Form"/>'s configuration.
+        /// </summary>
+        /// <typeparam name="TProperty">The type of the property for which a <see cref="FormField"/> will be added.</typeparam>
+        /// <param name="property">The property of <typeparamref name="T"/> for which a field will be added to the configuration.</param>
+        /// <param name="options">List of options that can be used for the field that will be created</param>
+        /// <param name="attributeOverrides">additional attributes to further customize the field</param>
+        /// <returns>The current instance.</returns>
         public FormBuilder<T> AddOptions<TProperty>(Expression<Func<T, TProperty>> property,
                                                     IEnumerable<FormFieldOption> options,
                                                     FormFieldAttributeOverrides attributeOverrides = null)
         {
             MemberExpression me = property.Body as MemberExpression;
-            FormField field = new() { Name = me.Member.Name, Type = FormFieldType.Array };
+            string memberName = me.Member.Name;
+            FormField field = new() {  Name = memberName, Label = memberName, Type = FormFieldType.Array };
             Option<FormFieldAttribute> optionalFormFieldAttribute = me.Member.GetCustomAttribute<FormFieldAttribute>()
                                                                              .SomeNotNull();
 
-            field.Options = options;
+#if NET5_0_OR_GREATER
+            field = field with
+            {
+                Options = options,
+#if NET6_0_OR_GREATER
+                MaxSize = options.TryGetNonEnumeratedCount(out int count) ? count : options.Count()
+#else
+                MaxSize = options.Count()
+#endif
 
-            UpdateAttributesField(property, field, optionalFormFieldAttribute, attributeOverrides.SomeNotNull());
+            };
+
+#else
+            field.Options = options;
+#endif
+
+            field = UpdateAttributesField(property, field, optionalFormFieldAttribute, attributeOverrides.SomeNotNull());
 
             _fields.Add(field);
 
             return this;
         }
 
+        /// <summary>
+        /// Adds a field to the <see cref="Form"/>'s configuration.
+        /// </summary>
+        /// <typeparam name="TProperty">The type of the property for which a <see cref="FormField"/> will be added.</typeparam>
+        /// <param name="property">The property of <typeparamref name="T"/> for which a field will be added to the configuration.</param>
+        /// <param name="options">List of options that can be used for the field that will be created. Each value will be used a <see cref="FormFieldOption.Label"/> and <see cref="FormFieldOption.Value"/> in resulting <see cref="FormFieldOption"/>.</param>
+        /// <param name="attributeOverrides">Overrides field's attributes</param>
+        /// <returns>The current instance.</returns>
         public FormBuilder<T> AddOptions<TProperty>(Expression<Func<T, TProperty>> property,
                                                     IEnumerable<string> options,
                                                     FormFieldAttributeOverrides attributeOverrides = null)
