@@ -1,158 +1,191 @@
 ﻿using FluentAssertions;
 
-using Newtonsoft.Json;
-
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
+using System.Text.Json;
 
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
-using Xunit.Sdk;
-using static Forms.FormFieldType;
-using static Newtonsoft.Json.JsonConvert;
 
-namespace Forms.UnitTests
+using static Candoumbe.Forms.FormFieldType;
+
+namespace Candoumbe.Forms.UnitTests;
+
+[UnitTest]
+public class FormBuilderTests(ITestOutputHelper outputHelper)
 {
-    [UnitTest]
-    public class FormBuilderTests
+    public class SuperHero
     {
-        private readonly ITestOutputHelper _outputHelper;
+        public Guid Id { get; init; }
+        
+        public string Nickname { get; set; }
 
-        public class SuperHero
+        [FormField(Secret = true)]
+        public string RealName { get; set; }
+
+        [FormField(MaxSize = 2)]
+        public IEnumerable<string> Cities { get; set; }
+
+        public DateOnly? BirthDate { get; set; }
+
+        public DateTime? LastBattleDate { get; set; }
+
+        public int? CurrentWinningStreakCount { get; set; }
+    }
+
+    public static TheoryData<FormBuilder<SuperHero>, Form> StronglyTypedBuilderCases
+        => new()
         {
-            public string Nickname { get; set; }
-
-            [FormField(Secret = true)]
-            public string RealName { get; set; }
-
-            [FormField(MaxSize = 2)]
-            public IEnumerable<string> Cities { get; set; }
-
-            public DateTime? BirthDate { get; set; }
-
-            public DateTime? LastBattleDate { get; set; }
-
-            public int? CurrentWinningStreakCount { get; set; }
-        }
-
-        public FormBuilderTests(ITestOutputHelper outputHelper)
-        {
-            _outputHelper = outputHelper;
-        }
-
-        public static IEnumerable<object[]> StronglyTypedBuilderCases
-        {
-            get
             {
-                yield return new object[]
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.Nickname),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.Nickname),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Count() == 1
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.Nickname)
-                            && field.Type == FormFieldType.String
-                            && field.Label == nameof(SuperHero.Nickname)
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.Nickname),
+                            Type = FormFieldType.String,
+                            Label = nameof(SuperHero.Nickname)
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.RealName),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.RealName),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Exactly(1)
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.RealName)
-                            && field.Type == FormFieldType.String
-                            && field.Label == nameof(SuperHero.RealName)
-                            && field.Secret.HasValue && field.Secret.Value
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.RealName),
+                            Type = FormFieldType.String,
+                            Label = nameof(SuperHero.RealName),
+                            Secret = true
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.RealName, new FormFieldAttributeOverrides { Secret = false }),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.RealName, new FormFieldAttributeOverrides { Secret = false }),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Count() == 1
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.RealName)
-                            && field.Type == FormFieldType.String
-                            && field.Label == nameof(SuperHero.RealName)
-                            && field.Secret.HasValue
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.RealName),
+                            Type = FormFieldType.String ,
+                            Label = nameof(SuperHero.RealName),
+                            Secret = false
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.RealName, new FormFieldAttributeOverrides { Description = "Secret identity of the hero" }),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.RealName, new FormFieldAttributeOverrides { Description = "Secret identity of the hero" }),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Count() == 1
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.RealName)
-                            && field.Type == FormFieldType.String
-                            && field.Label == nameof(SuperHero.RealName)
-                            && field.Description == "Secret identity of the hero"
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.RealName),
+                            Type = FormFieldType.String,
+                            Label = nameof(SuperHero.RealName),
+                            Secret = true,
+                            Description = "Secret identity of the hero"
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.BirthDate),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.BirthDate),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Count() == 1
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.BirthDate)
-                            && field.Type == FormFieldType.DateTime
-                            && field.Label == nameof(SuperHero.BirthDate)
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.BirthDate),
+                            Type = Date,
+                            Label = nameof(SuperHero.BirthDate)
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.CurrentWinningStreakCount),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddField(x => x.CurrentWinningStreakCount),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Count() == 1
-                        && form.Fields.Once(field => field.Name == nameof(SuperHero.CurrentWinningStreakCount)
-                                                     && field.Type == Integer
-                                                     && field.Label == nameof(SuperHero.CurrentWinningStreakCount)
-                        )
-                    )
-                };
-
-                yield return new object[]
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.CurrentWinningStreakCount),
+                            Type = Integer,
+                            Label = nameof(SuperHero.CurrentWinningStreakCount)
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddOptions(x => x.Cities, options: ["Metropolis", "Gotham", "Central city"]),
+                new Form
                 {
-                    new FormBuilder<SuperHero>()
-                        .AddOptions(x => x.Cities, options: new []{ "Metropolis", "Gotham", "Central city" }),
-                    (Expression<Func<Form, bool>>)(form => form.Fields != null && form.Fields.Exactly(1)
-                                                          && form.Fields.Once(field => field.Name == nameof(SuperHero.Cities)
-                                                                                       && field.Type == FormFieldType.Array
-                                                                                       && field.MinSize == null
-                                                                                       && field.MaxSize == 2
-                                                          )
-                    )
-                };
+                    Fields =
+                    [
+                        new FormField
+                        {
+                            Name = nameof(SuperHero.Cities),
+                            Type = FormFieldType.Array,
+                            Label = nameof(SuperHero.Cities),
+                            MaxSize = 2,
+                            Options =
+                            [
+                                new FormFieldOption("Metropolis", "Metropolis"),
+                                new FormFieldOption("Gotham", "Gotham"),
+                                new FormFieldOption("Central city", "Central city")
+                            ]
+                        } 
+                    ]
+                }
+            },
+            {
+                new FormBuilder<SuperHero>()
+                    .AddField(x => x.Id),
+                new Form
+                {
+                    Fields = [
+                        new FormField { Name = "Id", Type = FormFieldType.String, Label = nameof(SuperHero.Id) },
+                    ]
+                }
             }
-        }
+        };
 
-        [Theory]
-        [MemberData(nameof(StronglyTypedBuilderCases))]
-        public void StronglyTypedBuilder(FormBuilder<SuperHero> builder, Expression<Func<Form, bool>> formExpectation)
-        {
-            _outputHelper.WriteLine($"FormBuilder : {SerializeObject(builder, new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore })}");
+    [Theory]
+    [MemberData(nameof(StronglyTypedBuilderCases))]
+    public void StronglyTypedBuilder(FormBuilder<SuperHero> builder, Form expected)
+    {
+        outputHelper.WriteLine($"FormBuilder :  {builder.Jsonify(new JsonSerializerOptions(JsonSerializerDefaults.Web))}");
 
-            // Act
-            Form form = builder.Build();
-            _outputHelper.WriteLine($"Form built : {form}");
+        // Act
+        Form actual = builder.Build();
 
-            // Assert
-            form.Fields.Should()
-                       .NotBeNull().And
-                       .NotContainNulls();
-            form.Should()
-                .Match(formExpectation);
-        }
+        // Assert
+        actual.Fields.Should()
+            .NotBeNull().And
+            .NotContainNulls();
+        actual.Should().BeEquivalentTo(expected);
     }
 }
