@@ -1,11 +1,8 @@
-﻿namespace Forms
+﻿namespace Candoumbe.Forms
 {
     /// <summary>
     /// Link relation type representation (see http://www.iana.org/assignments/link-relations/link-relations.xhtml)
     /// </summary>
-    /// <remarks>
-    ///     
-    /// </remarks>
     public static class LinkRelation
     {
         /// <summary>
